requirements.txt
python-binance
websocket_client
schedule
